Solve graphql refetching bug (Issue #33)
The problem was, that apollo-client does not support cursor-based pagination and polling together.
The polling only updates the items that were loaded in the initial query (the first 10 items in our case),
and not does not poll posts that were added using fetchMore.
Because it's not possible to use a state for example,
to dinamically change what gets queried, we need to do this Mickey Mouse bs manually.
If we call stopPolling() and startPolling(), it will use the new values from the state
to run the polling. Thus until we load the next page, it will poll for the correct data.
When we load the next page, we override the data that needs to be polled.